requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.32
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
